complete: add extra-verbose support to _rg_types
When the extra-verbose style is set for the types tag, completed types are
displayed along with the patterns they correspond to. This can be enabled by
e.g. adding the following to .zshrc:

  zstyle ':completion:*:rg:*:types' extra-verbose true

This change also makes _rg_types use the actual rg specified on the command
line to look up types, and it fixes a mangled complete-all style check

Fixes #2195