Fix: strip standalone "." directory path
The problem is that in Ignore::matched, the perfix "./" is stripped from
files. In Ignore::matched_ignore, the directory path is then stripped
from file paths. Now imagine we have a hidden file "./.foo" and pass "."
as the search path. "./.foo" gets first stripped to ".foo" and then it
would have been stripped to "foo".