Add option to ignore nested git repositories
This implements the suggestion made in #23 to provide an option to ignore nested
git repositories.

A nested git repository is identified by the presence of a .git file or
directory. It's a directory in the regular case, but it's a file for git
worktrees and git submodules.

This option is disabled by default.