deps: Trust ryu, serde_json and syn
All of there crates are published by David Tolnay.  He is one of the
pillars of the rust ecosystem and I trust his release process enough for
now.  In the best case, ryo, serde_json and syn would be properly
audited, but it's a time cosuming undertaking I'm not willing to commit
to.