Separate flags from paths in decompression tool invocations
This fixes the bug when the filename of a compressed file starts with the `-` character and it gets passed to the decompression tool, but gets treated as an option instead of file.

In the best case, the option would be invalid and the decompression tool would finish with an error. In the worst case, the filename that looks like a valid option could potentially overwrite or remove data as a side effect.

All decompression tools in use here follow the convention that everything after `--` is treated as a filename.